Move time.svg to ./free-trial/assets/
